+++ test/unit/accessories/services/fanService.test.ts
@@ -337,7 +337,7 @@ describe('FanService', () => {
 
       expect(mockMqttClient.publishCommand).toHaveBeenCalledWith(
         expect.objectContaining({
-          data: { auto: 'ON', fmod: 'AUTO' },
+          data: { fpwr: 'ON', auto: 'ON', fmod: 'AUTO' },
         }),
       );
     });